Create Weinbergen
Hi,
i’m really not good at coding, but I’ve still tried my best because I really want to create my own map with a high resolution. If there are any errors in the code and it needs to be adjusted, I would kindly ask if you could make the adjustments. Thank you so much.